Change hosts file search
While the configuration directory `~/Library/Application Support/` may
be the proper place to write a configuration on macOS, we should be more
lenient what *reading* the configuration.

Search from a preset set of directories regardless of the type of the
host platform.  It is simpler and gets rid of a dependency.